although this should work for pgsql, it does not :shrug:

+++ src/Generators/Webserver/Database/Drivers/PostgreSQL.php
@@ -35,9 +35,7 @@ class PostgreSQL implements DatabaseGenerator
         $connection = $connection->system();
 
         $user = function () use ($connection, $config) {
-            $exists = $connection->statement("SELECT 1 FROM pg_roles WHERE rolname='{$config['username']}'");
-
-            if (!$exists && config('tenancy.db.auto-create-tenant-database-user')) {
+            if (config('tenancy.db.auto-create-tenant-database-user') && !$this->userExists($connection, $config['username']) ) {
                 return $connection->statement("CREATE USER \"{$config['username']}\" WITH PASSWORD '{$config['password']}'");
             }
 
@@ -51,6 +49,13 @@ class PostgreSQL implements DatabaseGenerator
         };
 
         return $user() && $create() && $grant();
+    }
+
+    protected function userExists($connection, string $username): bool
+    {
+        return $connection->table('pg_roles')
+            ->where('rolname', $username)
+            ->count() > 0;
     }
 
     /**
@@ -80,8 +85,10 @@ class PostgreSQL implements DatabaseGenerator
      */
     public function deleted(Deleted $event, array $config, Connection $connection): bool
     {
+        $connection->get()->disconnect();
+
         $user = function () use ($connection, $config) {
-            if (config('tenancy.db.auto-delete-tenant-database-user')) {
+            if (config('tenancy.db.auto-delete-tenant-database-user') && $this->userExists($connection->system(), $config['username'])) {
                 return $connection->system()->statement("DROP USER \"{$config['username']}\"");
             }
